requirements.txt
streamlit
tensorflow
segmentation-models==1.0.1
numpy
Pillow
opencv-python-headless
